+++ Changelog.md
@@ -2,6 +2,11 @@
 
 ## 0.2
 
+### 0.2.3 (??/??/??)
+
+- Add `Iterator <--> Fiber` interface.
+- API change: `into_inter()` takes a closure.
+
 ### 0.2.2 (18/07/2023)
 
 - Add example: [`monster.rs`](./examples/monster.rs)